move error handling to the rate limiter file
